refactor(desktop): extract library explorer state
