+++ apps/desktop/src/pages/library/model/libraryWorkspaceView.ts
@@ -1,0 +1,1 @@
+export type WorkspaceView = 'references' | 'records' | 'tag-settings' | 'preset-settings';
